added datacell component and changed data transformation
