+++ frontend/src/components/index.ts
@@ -2,3 +2,4 @@ export * from './dialog/Dialog';
 export * from './error/Error';
 export * from './loading/Loading';
 export * from './table/TableContainer';
+export * from './tablecells/DateCell';
